third_party/WebKit/LayoutTests/platform/win/external/wpt/dom/nodes/ProcessingInstruction-literal-1-expected.txt

This is a testharness.js-based test.
PASS <?xml?> is not a ProcessingInstruction 
Harness: the test ran to completion.
